+++ lib/Elastica/QueryBuilder/DSL/Query.php
@@ -2,6 +2,7 @@
 namespace Elastica\QueryBuilder\DSL;
 
 use Elastica\Exception\NotImplementedException;
+use Elastica\Query\AbstractSpanQuery;
 use Elastica\Query\BoolQuery;
 use Elastica\Query\Boosting;
 use Elastica\Query\Common;
@@ -26,11 +27,14 @@ use Elastica\Query\QueryString;
 use Elastica\Query\Range;
 use Elastica\Query\Regexp;
 use Elastica\Query\SimpleQueryString;
+use Elastica\Query\SpanContaining;
 use Elastica\Query\SpanFirst;
 use Elastica\Query\SpanMulti;
 use Elastica\Query\SpanNear;
+use Elastica\Query\SpanNot;
 use Elastica\Query\SpanOr;
 use Elastica\Query\SpanTerm;
+use Elastica\Query\SpanWithin;
 use Elastica\Query\Term;
 use Elastica\Query\Terms;
 use Elastica\Query\Type;
@@ -401,11 +405,16 @@ class Query implements DSL
     /**
      * span not query.
      *
+     * @param AbstractSpanQuery|null $include
+     * @param AbstractSpanQuery|null $exclude
+     *
+     * @return SpanNot
+     *
      * @link https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-span-not-query.html
      */
-    public function span_not()
+    public function span_not(AbstractSpanQuery $include = null, AbstractSpanQuery $exclude = null)
     {
-        throw new NotImplementedException();
+        return new SpanNot($include, $exclude);
     }
 
     /**
@@ -434,6 +443,36 @@ class Query implements DSL
     public function span_term(array $term = [])
     {
         return new SpanTerm($term);
+    }
+
+    /**
+     * span_containing query.
+     *
+     * @param AbstractSpanQuery|null $little
+     * @param AbstractSpanQuery|null $big
+     *
+     * @return SpanContaining
+     *
+     * @link https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-span-containing-query.html
+     */
+    public function span_containing(AbstractSpanQuery $little = null, AbstractSpanQuery $big = null)
+    {
+        return new SpanContaining($little, $big);
+    }
+
+    /**
+     * span_within query.
+     *
+     * @param AbstractSpanQuery|null $little
+     * @param AbstractSpanQuery|null $big
+     *
+     * @return SpanWithin
+     *
+     * @link https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-span-within-query.html
+     */
+    public function span_within(AbstractSpanQuery $little = null, AbstractSpanQuery $big = null)
+    {
+        return new SpanWithin($little, $big);
     }
 
     /**